requirements.txt
numpy==1.16.5
paceutils.egg==info
pandas==0.25.1
